Add messy Dockerfile for building twelf wasm dependencies
My goal is to set down once and for all a reproducible build process
that shows how the `twelf.wasm` was constructed, instead of it
existing ad hoc across my memory and comments on the
PR https://github.com/MLton/mlton/pull/550 . Already I discovered a
couple of crucial omissions in those comments.

This still deserves some cleanup, as the ordering of steps is not
really a coherent narrative, because I added some as I discovered
them. A cleaner one is forthcoming.